It helps if you refactor common fixtures too, Will
